+++ Sources/SyndiKit/Formats/Feeds/RSS/Enclosure.swift
@@ -15,21 +15,27 @@ public struct Enclosure: Codable {
     let container = try decoder.container(keyedBy: Self.CodingKeys.self)
     url = try container.decode(UTF8EncodedURL.self, forKey: .url).value
     type = try container.decode(String.self, forKey: .type)
+    length = try Self.decodeLength(from: container)
+  }
+
+  private static func decodeLength(
+    from container: KeyedDecodingContainer<CodingKeys>
+  ) throws -> Int? {
     if container.contains(.length) {
       do {
-        length = try container.decode(Int.self, forKey: .length)
+        return try container.decode(Int.self, forKey: .length)
       } catch {
         let lengthString = try container.decode(String.self, forKey: .length)
         if lengthString.isEmpty {
-          length = nil
+          return nil
         } else if let length = Int(lengthString) {
-          self.length = length
+          return length
         } else {
           throw error
         }
       }
     } else {
-      length = nil
+      return nil
     }
   }
 }
