+++ Sources/SyndiKit/Formats/Media/Podcast/PodcastLocation+GeoURI.swift
@@ -37,6 +37,7 @@ extension PodcastLocation {
       try? self.init(singleValue: description)
     }
 
+    // swiftlint:disable:next function_body_length
     public init(singleValue: String) throws {
       let pathComponents = try Self.pathComponents(from: singleValue)
 
@@ -46,10 +47,8 @@ extension PodcastLocation {
         let longitude = geoCoords[safe: 1]?.asDouble()
       else {
         throw DecodingError.dataCorrupted(
-          .init(
-            codingPath: [PodcastLocation.CodingKeys.geo],
-            debugDescription: "Invalid coordinates for geo attribute: \(singleValue)"
-          )
+          codingKey: PodcastLocation.CodingKeys.geo,
+          debugDescription: "Invalid coordinates for geo attribute: \(singleValue)"
         )
       }
 
@@ -80,18 +79,14 @@ extension PodcastLocation {
       guard
         components[safe: 0] == "geo" else {
         throw DecodingError.dataCorrupted(
-          .init(
-            codingPath: [PodcastLocation.CodingKeys.geo],
-            debugDescription: "Invalid prefix for geo attribute: \(string)"
-          )
+          codingKey: PodcastLocation.CodingKeys.geo,
+          debugDescription: "Invalid prefix for geo attribute: \(string)"
         )
       }
       guard let geoPath = components[safe: 1] else {
         throw DecodingError.dataCorrupted(
-          .init(
-            codingPath: [PodcastLocation.CodingKeys.geo],
-            debugDescription: "Invalid path for geo attribute: \(string)"
-          )
+          codingKey: PodcastLocation.CodingKeys.geo,
+          debugDescription: "Invalid path for geo attribute: \(string)"
         )
       }
