+++ Sources/SyndiKit/Formats/Media/Podcast/PodcastLocation+OsmQuery.swift
@@ -7,10 +7,6 @@ extension PodcastLocation {
       case node = "N"
       case way = "W"
       case relation = "R"
-
-      internal static func isValid(_ rawValue: String) -> Bool {
-        OsmType(rawValue: rawValue) != nil
-      }
     }
 
     public let id: Int
@@ -24,18 +20,14 @@ extension PodcastLocation {
 
       guard let osmType = osmStr.removeFirst().asOsmType() else {
         throw DecodingError.dataCorrupted(
-          .init(
-            codingPath: [PodcastLocation.CodingKeys.osmQuery],
-            debugDescription: "Invalid type for osm attribute: \(osmStr)"
-          )
+          codingKey: PodcastLocation.CodingKeys.osmQuery,
+          debugDescription: "Invalid type for osm attribute: \(osmStr)"
         )
       }
       guard let osmID = osmStr.split(separator: "#")[safe: 0]?.asExactInt() else {
         throw DecodingError.dataCorrupted(
-          .init(
-            codingPath: [PodcastLocation.CodingKeys.osmQuery],
-            debugDescription: "Invalid id of type Int for osm attribute: \(osmStr)"
-          )
+          codingKey: PodcastLocation.CodingKeys.osmQuery,
+          debugDescription: "Invalid id of type Int for osm attribute: \(osmStr)"
         )
       }
       let osmRevision = osmStr.split(separator: "#")[safe: 1]?.asInt()
